requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-socketio>=5.11
twitchio>=3.2
websockets>=13.0
pydantic>=2.0
pyyaml>=6.0
httpx>=0.27
